+++ .gitignore
@@ -129,4 +129,4 @@ dmypy.json
 .pyre/
 
 # Gitpod
-# /.theia
+/.theia
