+++ README.md
@@ -1,3 +1,3 @@
-A library for calculating optimal routes for each train (by revenue) for the board game "1846: The Race For The Midwest".
+A library for calculating optimal routes for each train (by revenue) for the 18xx family of board games.
 
-Usage instructions forthcoming. For now, check out the [related web app](https://routes1846.herokuapp.com/).
+Usage instructions forthcoming.
