+++ routes18xx/tiles.py
@@ -1,7 +1,7 @@
 import collections
 import json
 
-from routes1846 import get_data_file
+from routes18xx import get_data_file
 
 
 _TILE_FILENAME = "tiles.json"
@@ -44,7 +44,7 @@ class Tile(object):
 
 
 def _load_all():
-    with open(get_data_file(_TILE_FILENAME)) as tiles_file:
+    with open(get_data_file(game, _TILE_FILENAME)) as tiles_file:
         tiles_json = json.load(tiles_file)
 
     return {int(id): Tile.create(int(id), **args) for id, args in tiles_json.items()}
